refactoring

+++ src/Cuda/CudaHelper.cu
@@ -13,13 +13,13 @@ void
 CudaHelper::calculateBlockThreadNum(int& threads_x, int& threads_y, int& blocks_x, int& blocks_y, int height,
                                     int width) {
     threads_x = (width >= THREAD_PER_TWO_DIM_BLOCK) ? THREAD_PER_TWO_DIM_BLOCK : width;
-    blocks_x = (int) ceil(1.0 * width / threads_x);
+    blocks_x = ceil(1.0 * width / threads_x);
     threads_y = (height >= THREAD_PER_TWO_DIM_BLOCK) ? THREAD_PER_TWO_DIM_BLOCK : height;
-    blocks_y = (int) ceil(1.0 * height / threads_y);
+    blocks_y = ceil(1.0 * height / threads_y);
 }
 
 void CudaHelper::allocateGpuMemory(float** data, int size) {
-    int bytes = size * sizeof(float);
+    int bytes = static_cast<int>(size * sizeof(float));
     cudaMalloc(data, bytes);
 }
 
@@ -28,16 +28,16 @@ void CudaHelper::deleteGpuMemory(float* data) {
 }
 
 void CudaHelper::copyFromCpuToGpu(float* cpuData, float* gpuData, int size) {
-    int bytes = size * sizeof(float);
+    int bytes = static_cast<int>(size * sizeof(float));
     cudaMemcpy(gpuData, cpuData, bytes, cudaMemcpyHostToDevice);
 }
 
 void CudaHelper::copyFromGpuToCpu(float* gpuData, float* cpuData, int size) {
-    int bytes = size * sizeof(float);
+    int bytes = static_cast<int>(size * sizeof(float));
     cudaMemcpy(cpuData, gpuData, bytes, cudaMemcpyDeviceToHost);
 }
 
 void CudaHelper::copyFromGpuToGpu(float* src, float* dest, int size) {
-    int bytes = size * sizeof(float);
+    int bytes = static_cast<int>(size * sizeof(float));
     cudaMemcpy(dest, src, bytes, cudaMemcpyDeviceToDevice);
 }
