+++ src/Model/CostFunction/BinaryCrossEntropy.cpp
@@ -7,7 +7,7 @@
 float BinaryCrossEntropy::calculate(const Matrix& YHat, const Matrix& Y) {
     Matrix first = Y * Matrix::log(YHat);
     Matrix second = (-Y + 1) * Matrix::log(-YHat + 1);
-    return -(first + second).sum() / Y.getWidth();
+    return -(first + second).sum() / static_cast<float>(Y.getWidth());
 }
 
 Matrix BinaryCrossEntropy::derivative(const Matrix& YHat, const Matrix& Y) {
