+++ src/Model/Layer/Linear.cpp
@@ -49,8 +49,8 @@ Matrix Linear::forwardWithCache(const Matrix& input) {
 
 Matrix Linear::backward(const Matrix& input, int m, float lr) {
     Matrix dx = activation->derivative(cache[1], input);
-    Matrix dW = Matrix::multiply(dx, cache[0]) / m;
-    Matrix db = Matrix::sum(dx, 0) / m;
+    Matrix dW = Matrix::multiply(dx, cache[0]) / static_cast<float>(m);
+    Matrix db = Matrix::sum(dx, 0) / static_cast<float>(m);
     Matrix output = Matrix::multiply(Matrix::transpose(W), dx);
     updateParams(dW, db, lr);
     return output;
@@ -74,7 +74,7 @@ void Linear::initWeights(int previousHidden) {
 }
 
 void Linear::serialize(std::ofstream& file) {
-    if(Config::getInstance().getProvider() == Provider::GPU){
+    if (Config::getInstance().getProvider() == Provider::GPU) {
         W.copyGpuToCpu();
         b.copyGpuToCpu();
     }
@@ -110,7 +110,7 @@ void Linear::deserialize(std::ifstream& file) {
         }
         file >> b[i][0];
     }
-    if(Config::getInstance().getProvider() == Provider::GPU){
+    if (Config::getInstance().getProvider() == Provider::GPU) {
         W.moveCpuToGpu();
         b.moveCpuToGpu();
     }
